Move the hook commands to their own package.
